LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-functions/circle-function-valid-expected.txt


PASS e.style['shape-outside'] = "circle()" should set the property value
PASS e.style['shape-outside'] = "circle(1px)" should set the property value
FAIL e.style['shape-outside'] = "circle(20px at center)" should set the property value assert_equals: serialization should be canonical expected "circle(20px at 50% 50%)" but got "circle(20px at center center)"
PASS e.style['shape-outside'] = "circle(at 10% 20%)" should set the property value
FAIL e.style['shape-outside'] = "circle(4% at top right)" should set the property value assert_equals: serialization should be canonical expected "circle(4% at 100% 0%)" but got "circle(4% at right top)"
PASS e.style['shape-outside'] = "circle(calc(100% - 20px) at calc(100% - 20px) calc(100% / 4))" should set the property value
FAIL e.style['shape-outside'] = "circle(closest-corner at center)" should set the property value assert_equals: serialization should be canonical expected "circle(closest-corner at 50% 50%)" but got "circle(closest-corner at center center)"
PASS e.style['shape-outside'] = "circle(closest-corner at 20px 50px)" should set the property value
FAIL e.style['shape-outside'] = "circle(closest-side at center)" should set the property value assert_equals: serialization should be canonical expected "circle(at 50% 50%)" but got "circle(at center center)"
PASS e.style['shape-outside'] = "circle(closest-side at 20px 30%)" should set the property value
FAIL e.style['shape-outside'] = "circle(farthest-corner at center top)" should set the property value assert_equals: serialization should be canonical expected "circle(farthest-corner at 50% 0%)" but got "circle(farthest-corner at center top)"
FAIL e.style['shape-outside'] = "circle(farthest-corner at center)" should set the property value assert_equals: serialization should be canonical expected "circle(farthest-corner at 50% 50%)" but got "circle(farthest-corner at center center)"
FAIL e.style['shape-outside'] = "circle(farthest-side at center top)" should set the property value assert_equals: serialization should be canonical expected "circle(farthest-side at 50% 0%)" but got "circle(farthest-side at center top)"
FAIL e.style['shape-outside'] = "circle(farthest-side at center)" should set the property value assert_equals: serialization should be canonical expected "circle(farthest-side at 50% 50%)" but got "circle(farthest-side at center center)"
